+++ src/main/kotlin/database/DataBase.kt
@@ -13,16 +13,16 @@ const val password = "c0mp0ly"
 const val EMPTY_HISTORY_TEXT = ""
 
 object UserScore : Table() {
-    val userId = long("user_id")
-    val chatId = long("chat_id")
+    val userId = integer("user_id")
+    val chatId = integer("chat_id")
     val reputation = integer("reputation")
     val history_respects = text("history_respects")
     val history_disrespects = text("history_disrespects")
 }
 
 object UserReward : Table() {
-    val userId = long("user_id")
-    val chatId = long("chat_id")
+    val userId = integer("user_id")
+    val chatId = integer("chat_id")
     val rewardName = text("reward_name")
 }
